Remove can bus Recorder
